tools/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
openai
anthropic
google-search-results
requests
beautifulsoup4
pdf2image
Pillow
pydantic
